re-using filter components between events and courses
